+++ src/Scrabble.Domain/Model/Tile.cs
@@ -2,9 +2,9 @@
 
 namespace Scrabble.Domain.Model
 {
-    public class Tile(Char letter, int Value = 1)
+    public class Tile(Char letter, ushort Value = 1)
     {
         public Char Letter { get; } = Char.ToUpper(letter);
-        public int Value { get; } = Value;
+        public ushort Value { get; } = Value;
     }
 }
